Update pill sorting
