+++ src/Components/PillList/PillList.jsx
@@ -2,19 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { PILL_ITEM_ARRAY } from '../../Constants/PropTypes';
 import Pill from '../Pill/Pill';
+import { pillSort } from '../../utilities';
 
 const PillList = ({ items, onPillClick }) => (
   <div>
     {
-      (items.sort((a, b) => { // sort our pills by description or code
-        const descA = (a.description || a.code).toLowerCase();
-        const descB = (b.description || b.code).toLowerCase();
-        if (descA < descB) { // sort string ascending
-          return -1;
-        }
-        if (descA > descB) { return 1; }
-        return 0; // default return value (no sorting)
-      }))
+      (items.slice().sort(pillSort))
         .map(item =>
           (<Pill
             key={`${item.codeRef}-${item.description}`}
