+++ src/Components/SearchFilters/SearchFiltersContainer/SearchFiltersContainer.jsx
@@ -5,7 +5,7 @@ import MultiSelectFilter from '../MultiSelectFilter/MultiSelectFilter';
 import BooleanFilterContainer from '../BooleanFilterContainer/BooleanFilterContainer';
 import LanguageFilter from '../LanguageFilter/LanguageFilter';
 import { FILTER_ITEMS_ARRAY, ACCORDION_SELECTION_OBJECT } from '../../../Constants/PropTypes';
-import { titleSort, descriptionSort } from './descriptionSort';
+import { titleSort, descriptionSort } from '../../../utilities';
 
 class SearchFiltersContainer extends Component {
   onBooleanFilterClick(isChecked, code, selectionRef) {
