+++ src/utilities.test.js
@@ -5,6 +5,9 @@ import { ajax,
          localStorageFetchValue,
          localStorageToggleValue,
          fetchUserToken,
+         descriptionSort,
+         titleSort,
+         pillSort,
        } from './utilities';
 
 const posts = [
@@ -94,5 +97,28 @@ describe('fetchUserToken', () => {
     const output = fetchUserToken();
     expect(output).toBe('Token 1234');
     localStorage.clear();
+  });
+});
+
+describe('sort functions', () => {
+  const items = [{ title: 'a', description: 'a' }, { title: 'b', description: 'b' }];
+  const pills = [{ description: 'a' }, { code: 'b' }];
+
+  it('can sort by description', () => {
+    expect(descriptionSort(items[0], items[1])).toBe(-1);
+    expect(descriptionSort(items[1], items[0])).toBe(1);
+    expect(descriptionSort(items[0], items[0])).toBe(0);
+  });
+
+  it('can sort by title', () => {
+    expect(titleSort(items[0], items[1])).toBe(-1);
+    expect(titleSort(items[1], items[0])).toBe(1);
+    expect(titleSort(items[0], items[0])).toBe(0);
+  });
+
+  it('can sort by description or code', () => {
+    expect(pillSort(pills[0], pills[1])).toBe(-1);
+    expect(pillSort(pills[1], pills[0])).toBe(1);
+    expect(pillSort(pills[0], pills[0])).toBe(0);
   });
 });
